// multi/src/app/model/shift.ts
/**********************************************************************************
* File Name   :   Shift.ts
* Description :   Shift
*
* Copyright (c) MultiSight™ © 2016 Schneider Electric 
* --------------------------------------------------------------------------------
* Date             Author              Comment
* --------------------------------------------------------------------------------
* 1-Nov-2016      QuEST Team          Initial version created
**********************************************************************************/

export class Shift {
 name: string;
 start_time: string;
 end_time: string;
  

}